Update version for loading dps dpp
