+++ scripts/tests/test_assign_custom_color_to_tractogram.py
@@ -13,7 +13,8 @@ tmp_dir = tempfile.TemporaryDirectory()
 
 
 def test_help_option(script_runner):
-    ret = script_runner.run('scil_assign_color_to_tractogram.py', '--help')
+    ret = script_runner.run('scil_assign_custom_color_to_tractogram.py',
+                            '--help')
     assert ret.success
 
 
@@ -21,7 +22,9 @@ def test_execution_tractometry(script_runner):
     os.chdir(os.path.expanduser(tmp_dir.name))
     in_bundle = os.path.join(get_home(), 'tractometry',
                              'IFGWM.trk')
-    ret = script_runner.run('scil_assign_color_to_tractogram.py',
-                            in_bundle, '--fill_color', '0x000000',
-                            '--out_tractogram', 'colored.trk')
+    in_anat = os.path.join(get_home(), 'tractometry',
+                           'IFGWM_labels_map.nii.gz')
+    ret = script_runner.run('scil_assign_custom_color_to_tractogram.py',
+                            in_bundle, 'colored.trk', '--from_anatomy',
+                            in_anat)
     assert ret.success
