+++ scripts/tests/test_assign_uniform_color_to_tractogram.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+import os
+import tempfile
+
+from scilpy.io.fetcher import get_testing_files_dict, fetch_data, get_home
+
+
+# If they already exist, this only takes 5 seconds (check md5sum)
+fetch_data(get_testing_files_dict(), keys=['tractometry.zip'])
+tmp_dir = tempfile.TemporaryDirectory()
+
+
+def test_help_option(script_runner):
+    ret = script_runner.run('scil_assign_uniform_color_to_tractogram.py',
+                            '--help')
+    assert ret.success
+
+
+def test_execution_tractometry(script_runner):
+    os.chdir(os.path.expanduser(tmp_dir.name))
+    in_bundle = os.path.join(get_home(), 'tractometry',
+                             'IFGWM.trk')
+    ret = script_runner.run('scil_assign_uniform_color_to_tractogram.py',
+                            in_bundle, '--fill_color', '0x000000',
+                            '--out_tractogram', 'colored.trk')
+    assert ret.success
